Store computed digests: Only store inside coursier cache

+++ modules/cache/jvm/src/main/scala/coursier/cache/FileCache.scala
@@ -681,7 +681,7 @@ import scala.util.control.NonFatal
                 Left(new ArtifactError.ChecksumFormatError(sumType, sumFile.getPath))
 
               case Some(sum) =>
-                val calculatedSum: BigInteger = FileCache.persistedDigest(sumType, localFile0)
+                val calculatedSum: BigInteger = FileCache.persistedDigest(location, sumType, localFile0)
 
                 if (sum == calculatedSum)
                   Right(())
@@ -1083,35 +1083,44 @@ object FileCache {
   private val checksumHeader = Seq("MD5", "SHA1", "SHA256")
 
   /**
-    * Store computed cache in files so we don't have to recompute them over and over.
+    * Store computed cache in a file so we don't have to recompute them over and over.
     */
-  private def persistedDigest(sumType: String, localFile: File): BigInteger = {
-    val cacheFile = auxiliaryFile(localFile, sumType + ".computed")
-    val cacheFilePath = cacheFile.toPath
+  private def persistedDigest(location: File, sumType: String, localFile: File): BigInteger = {
+    // only store computed files within coursier cache folder
+    val isInCache: Boolean = {
+      val location0 = location.getCanonicalPath.stripSuffix("/") + "/"
+      localFile.getCanonicalPath.startsWith(location0)
+    }
 
     val digested: Array[Byte] =
-      try Files.readAllBytes(cacheFilePath) catch {
-        case _: NoSuchFileException =>
-          val bytes: Array[Byte] = digest(sumType, localFile)
+      if (!isInCache) computeDigest(sumType, localFile)
+      else {
+        val cacheFile = auxiliaryFile(localFile, sumType + ".computed")
+        val cacheFilePath = cacheFile.toPath
 
-          // Atomically write file by using a temp file in the same directory
-          val tmpFile = File.createTempFile(cacheFile.getName, ".tmp", cacheFile.getParentFile).toPath
-          try {
-            Files.write(tmpFile, bytes)
-            try Files.move(tmpFile, cacheFilePath, StandardCopyOption.ATOMIC_MOVE)
-            catch {
-              // In the case of multiple processes/threads which all compute this digest, first thread wins.
-              case _: FileAlreadyExistsException => ()
-            }
-          } finally Files.deleteIfExists(tmpFile)
+        try Files.readAllBytes(cacheFilePath) catch {
+          case _: NoSuchFileException =>
+            val bytes: Array[Byte] = computeDigest(sumType, localFile)
 
-          bytes
+            // Atomically write file by using a temp file in the same directory
+            val tmpFile = File.createTempFile(cacheFile.getName, ".tmp", cacheFile.getParentFile).toPath
+            try {
+              Files.write(tmpFile, bytes)
+              try Files.move(tmpFile, cacheFilePath, StandardCopyOption.ATOMIC_MOVE)
+              catch {
+                // In the case of multiple processes/threads which all compute this digest, first thread wins.
+                case _: FileAlreadyExistsException => ()
+              }
+            } finally Files.deleteIfExists(tmpFile)
+
+            bytes
+        }
       }
 
     new BigInteger(1, digested)
   }
 
-  private def digest(sumType: String, localFile: File): Array[Byte] = {
+  private def computeDigest(sumType: String, localFile: File): Array[Byte] = {
     val md = MessageDigest.getInstance(sumType)
 
     var is: FileInputStream = null
